rework Review, User and ShoppingCart Controllers

+++ src/main/java/com/example/emag/model/dao/IProductDAO.java
@@ -1,6 +1,7 @@
 package com.example.emag.model.dao;
 
 import com.example.emag.model.pojo.Product;
+import com.example.emag.model.pojo.Specification;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -17,11 +18,11 @@ public interface IProductDAO {
     void editProduct(Product product) throws SQLException;// ADMIN STUFF
     void updateQuantity(Integer productId, Integer quantity) throws SQLException; //ADMIN STUFF
     void setDiscount(Integer productId, Integer discount) throws SQLException; //ADMIN STUFF
-    List<Product> getFavouriteProducts(Integer userId) throws SQLException;
-    Map<Product, Integer> getProductsByOrder(Integer orderId) throws SQLException;
+    List<Product> getFavouriteProducts(long userId) throws SQLException;
+    Map<Product, Integer> getProductsByOrder(long orderId) throws SQLException;
     //List<Order> getOrdersByUserId(Integer userId) throws SQLException;
-    void addFavouriteProduct(Integer userId, Integer productId) throws SQLException;
-    void removeFavouriteProduct(Integer userId, Integer productId) throws SQLException;
+    void addFavouriteProduct(long userId, long productId) throws SQLException;
+    void removeFavouriteProduct(long userId, long productId) throws SQLException;
     void addProductsToOrder(Map<Product, Integer> products, long orderId) throws SQLException;
     boolean checkIfProductsExist(Map<Product, Integer> products) throws SQLException;
     void removeProducts(Map<Product, Integer> products) throws SQLException;
